requirements.txt
pandas==2.1.3
graphviz==0.20.1
cplex==22.1.1.0
rdflib-neo4j==1.0.0
SPARQLWrapper==2.0.0
rdflib==7.0.0
python-dotenv==1.0.1
pyfactxx
owlready2==0.46
pyyaml==6.0.1
requests==2.23.0
pytest==8.2.0